Split `qlipp_compute` into `phantoms` and `reconstructions`
