+++ examples/3D-fluor-to-3D-density.py
@@ -1,10 +1,10 @@
 from waveorder.io.reader import WaveorderReader
 from waveorder.io.writer import WaveorderWriter
-from recOrder.compute.qlipp_compute import (
+from recOrder.compute.reconstructions import (
     initialize_reconstructor,
-    fluorescence_from_phantom,
     reconstruct_density_from_fluorescence,
 )
+from recOrder.compute.phantoms import fluorescence_from_phantom
 from datetime import datetime
 import numpy as np
 import napari
