+++ examples/3D-pol-to-birefringence.py
@@ -1,13 +1,13 @@
 from waveorder.io.reader import WaveorderReader
 from waveorder.io.writer import WaveorderWriter
 from recOrder.io.utils import load_bg
-from recOrder.compute.qlipp_compute import (
+from recOrder.compute.reconstructions import (
     initialize_reconstructor,
     reconstruct_qlipp_stokes,
     reconstruct_qlipp_birefringence,
     reconstruct_phase3D,
-    pol_3D_from_phantom,
 )
+from recOrder.compute.phantoms import pol_3D_from_phantom
 from datetime import datetime
 import numpy as np
 import napari
